refactor(decks): extract DeckCanvasViewModel to manage grid slots
